Fix defects surfaced by modern compiler warnings
Building against a current SDK turns up a number of genuine bugs that
were hidden when the project was last touched in 2011:

  - AppController: the PayPal donation URL was assembled with a format
    string containing a bare "%2egif". printf read %2e as a conversion
    and consumed an argument that was never passed, corrupting the
    trailing bn= parameter. Escape it as "%%2egif".
  - AppController: representationOfImageRepsInArray:usingType:
    properties: was passed nil for a non-null parameter.
  - PrefController: dealloc called [super release] instead of
    [super dealloc].
  - ControlUIView: -isDragging was sent to an NSCell, where it is not
    declared, so the result was typed as id. Cast the cell to
    TimeSliderCell, which does declare it.
  - RootLayerView: fabsf() was applied to CGFloat values, which are
    double on 64-bit, truncating them to float. Use fabs().
  - CocoaAppendix: a caller-supplied string was passed to
    NSGetAlertPanel as the format string.
  - InspectorController, ParameterManager, PlayListController: cast
    NSInteger/NSUInteger values used as %d arguments.

The remaining warnings are almost entirely renamed AppKit enum
constants (NSCommandKeyMask and friends). Those are left alone here so
this change stays reviewable; GCC_TREAT_WARNINGS_AS_ERRORS is turned
off in the meantime, as a 2011 codebase cannot be warning-clean
against a 2026 SDK.

+++ MPlayerX/InspectorController.m
@@ -171,9 +171,9 @@ static BOOL init_ed = NO;
 			if (mask & kLoadMIMaskTrackInfo) {
 				NSMutableArray *tracks = [[NSMutableArray alloc] initWithCapacity:3];
 				
-				cnt = [[mi videoInfo] count]; if (cnt) { [tracks addObject:[NSString stringWithFormat:kMPXStringInfoTrackInfoVideo, cnt]]; }
-				cnt = [[mi audioInfo] count]; if (cnt) { [tracks addObject:[NSString stringWithFormat:kMPXStringInfoTrackInfoAudio, cnt]]; }
-				cnt = [[mi subInfo] count];   if (cnt) { [tracks addObject:[NSString stringWithFormat:kMPXStringInfoTrackInfoSubtitle, cnt]]; }
+				cnt = [[mi videoInfo] count]; if (cnt) { [tracks addObject:[NSString stringWithFormat:kMPXStringInfoTrackInfoVideo, (int)cnt]]; }
+				cnt = [[mi audioInfo] count]; if (cnt) { [tracks addObject:[NSString stringWithFormat:kMPXStringInfoTrackInfoAudio, (int)cnt]]; }
+				cnt = [[mi subInfo] count];   if (cnt) { [tracks addObject:[NSString stringWithFormat:kMPXStringInfoTrackInfoSubtitle, (int)cnt]]; }
 				[trackInfo setStringValue:[[tracks componentsJoinedByString:@", "] stringByAppendingString:kMPXStringInfoTrackTrackText]];
 				
 				[tracks release];
